correct collisions with existing blocks

+++ JEWels/Constants.h
@@ -7,7 +7,7 @@ namespace Constants
 	const sf::Vector2f TOPLEFT(230, 65);
 	const sf::Vector2f BOARDSIZE(520, 680);
 	const sf::Vector2f BOTTOMRIGHT = TOPLEFT + BOARDSIZE;
-	const int GEMSIZE = 40;
+	const float GEMSIZE = 40;
 	const int ROWS = 13;
 	const int COLS = 17;
 	const float STEP = 40.0f;
